chore: move API views to separate files + run black formatter over all views - no real code logic changes
